backend/pyproject.toml
[tool.poetry]
name = "paroquianews"
version = "0.1.0"
description = "Sistema de gerenciamento de notícias e fotos para paróquias"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
django = "^5.0"
djangorestframework = "^3.14"
django-cors-headers = "^4.3"
pillow = "^10.1.0"
python-dotenv = "^1.0.0"
django-filter = "^23.5"
django-storages = "^1.14.2"
psycopg2-binary = "^2.9.9"  # Para PostgreSQL
djangorestframework-simplejwt = "^5.3.1"  # Para autenticação JWT

[tool.poetry.group.dev.dependencies]
pytest = "^7.4"
pytest-django = "^4.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 